# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Mobile App Development - 2019",
      "license": {
        "name": "MIT"
      }
    },
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authToken"
      }
    },
    "paths": {
        "/auth": {
            "post": {
                "summary": "Authenticate a user",

                "tags": [
                  "Authentication"
                ],
                "parameters": [
                  {
                    "in": "body",
                    "name": "body",
                    "description": "Authentication object",
                    "required": true,
                    "schema": {
                      "$ref": "#/definitions/Authentication"
                    }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "Null response"
                  },
                  "default": {
                    "description": "unexpected error",
                    "schema": {
                      "$ref": "#/definitions/Error"
                    }
                  }
                }
              }
          },
      "/contacts": {
        "get": {
          "summary": "List all contacts",
          "tags": [
            "Contacts"
          ],
          "security": [
            {
              "ApiKeyAuth": [
                "authToken"
              ]
            }
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "An paged array of contacts",
              "headers": {
                "x-next": {
                  "type": "string",
                  "description": "A link to the next page of responses"
                }
              },
              "schema": {
                "$ref": "#/definitions/Contacts"
              }
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "post": {
          "summary": "Create a contact",
          "tags": [
            "Contacts"
          ],
          "security": [
            {
              "ApiKeyAuth": [
                "authToken"
              ]
            }
          ],
          "parameters": [
            {
              "in": "body",
              "name": "contact",
              "description": "The contact to create.",
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "contactNumber": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Null response"
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/contacts/{contactId}": {
        "get": {
          "summary": "Info for a specific contact",
          "tags": [
            "Contacts"
          ],
          "security": [
            {
              "ApiKeyAuth": [
                "authToken"
              ]
            }
          ],
          "parameters": [
            {
              "name": "contactId",
              "in": "path",
              "required": true,
              "description": "The id of the contact to retrieve",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Expected response to a valid request",
              "schema": {
                "$ref": "#/definitions/Contacts"
              }
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "put": {
            "summary": "Update a specific contact",
            "tags": [
              "Contacts"
            ],
            "security": [
              {
                "ApiKeyAuth": [
                  "authToken"
                ]
              }
            ],
            "parameters": [
              {
                "name": "contactId",
                "in": "path",
                "required": true,
                "description": "The id of the contact to retrieve",
                "type": "string"
              },
              {
                "in": "body",
                "name": "contact",
                "description": "The contact to update.",
                "schema": {
                  "type": "object",
                  "properties": {
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "contactNumber": {
                      "type": "string"
                    }
                  }
                }
              }
            ],
            "responses": {
              "200": {
                "description": "Expected response to a valid request",
                "schema": {
                  "$ref": "#/definitions/Contacts"
                }
              },
              "default": {
                "description": "unexpected error",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
        },
        "delete": {
            "summary": "Delete a specific contact",
            "operationId": "deleteContactById",
            "tags": [
              "Contacts"
            ],
            "security": [
              {
                "ApiKeyAuth": [
                  "authToken"
                ]
              }
            ],
            "parameters": [
              {
                "name": "contactId",
                "in": "path",
                "required": true,
                "description": "The id of the contact to retrieve",
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "Expected response to a valid request",
                "schema": {
                  "$ref": "#/definitions/Contacts"
                }
              },
              "default": {
                "description": "unexpected error",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
        }
      }
    },
    "definitions": {
      "Contact": {
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string"
          }
        }
      },
      "Contacts": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Contact"
        }
      },
      "Authentication": {
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Error": {
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }